Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webrtc/receiver-track-live.https-expected.txt



PASS Setup audio call
FAIL Inactivate the audio transceiver assert_equals: expected "live" but got "ended"
FAIL Reactivate the audio transceiver assert_equals: expected "live" but got "ended"
PASS Clean-up
